requirements.txt
numpy >= 1.18.4
PyYAML >= 5.3.1
cxxfilt >= 0.2.0
tqdm >= 4.47.0
torch >= 1.5.1
torchvision >= 0.6.1
